pyproject.toml
[project]
name = "ml-learning"
version = "0.1.0"
description = "Multi-framework machine learning practice repository"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    # Workspace packages
    "pytorch-learning",
    "pyro-learning",
    
    # Data manipulation & scientific computing
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "scipy>=1.12.0",
    
    # Visualization
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "plotly>=5.18.0",
    
    # Jupyter notebooks
    "jupyterlab>=4.0.0",
    "ipykernel>=6.29.0",
    "ipywidgets>=8.1.0",
    
    # CLI tooling
    "typer>=0.9.0",
    "rich>=13.0.0",
    
    # Utilities
    "tqdm>=4.66.0",
]

[dependency-groups]
dev = [
    "ruff>=0.2.0",
    "ty",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]

[tool.uv.workspace]
members = ["packages/*"]

[tool.uv.sources]
pytorch-learning = { workspace = true }
pyro-learning = { workspace = true }

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "F401",  # unused imports (useful during development)
]

[tool.ruff.format]
skip-magic-trailing-comma = false

[tool.ty]
src = ["packages/*/src"]
python-version = "3.11"

[tool.pytest.ini_options]
testpaths = ["packages/*/tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"
